use immediate window in persoal-settings
